# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ace Digital OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List accessible channels
      responses:
        "200":
          description: Channel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          description: Invalid input
        "403":
          description: Forbidden
  /v1/channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get channel details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update channel (owners / admins)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete channel (super admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /v1/channels/{id}/members:
    get:
      operationId: listChannelMembers
      tags: [channels]
      summary: List channel members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelMember"
    post:
      operationId: addChannelMember
      tags: [channels]
      summary: Add member to channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelMember"
        "400":
          description: Invalid input
        "403":
          description: Forbidden
  /v1/channels/{id}/members/{userId}:
    delete:
      operationId: removeChannelMember
      tags: [channels]
      summary: Remove member from channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "400":
          description: Cannot remove last owner
        "403":
          description: Forbidden
        "404":
          description: Not found
  /v1/channels/{id}/messages:
    get:
      operationId: getChannelMessages
      tags: [channels]
      summary: Get channel messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [channels]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /v1/channels/{id}/messages/{messageId}/poll/vote:
    patch:
      operationId: votePoll
      tags: [channels]
      summary: Vote on a poll message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollVoteInput"
      responses:
        "200":
          description: Updated poll message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /v1/channels/{id}/messages/{messageId}/event/rsvp:
    patch:
      operationId: rsvpEvent
      tags: [channels]
      summary: RSVP to an event message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRsvpInput"
      responses:
        "200":
          description: Updated event message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ACTIVITY
...
components:
  schemas:
...
    Channel:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        type:
          type: string
        visibility:
          type: string
        archived:
          type: boolean
        memberCount:
          type: integer
        myRole:
          type: ["string", "null"]
        unreadCount:
          type: integer
        createdAt:
          type: string
    ChannelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [TEAM, ANNOUNCEMENT]
        teamId:
          type: integer
        memberIds:
          type: array
          items:
            type: integer
    ChannelUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        archived:
          type: boolean
    ChannelMember:
      type: object
      required: [userId, fullName, role]
      properties:
        userId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        joinedAt:
          type: string
    ChannelMemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
        role:
          type: string
          enum: [owner, member, viewer]
    MessageAttachment:
      type: object
      required: [type, url]
      properties:
        type:
          type: string
          enum: [image, file, video, audio]
        url:
          type: string
        name:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        thumbUrl:
          type: string
    Message:
      type: object
      required: [id, channelId, senderId, body, createdAt]
      properties:
        id:
          type: integer
        channelId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderAvatar:
          type: ["string", "null"]
        body:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/MessageAttachment"
        messageKind:
          type: string
          enum: [text, poll, event]
        metadata:
          type: object
        createdAt:
          type: string
    MessageInput:
      type: object
      properties:
        body:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/MessageAttachment"
        messageKind:
          type: string
          enum: [text, poll, event]
        metadata:
          type: object
    PollVoteInput:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: string
    EventRsvpInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [going, maybe, no]

    # ACTIVITY & NOTIFICATIONS
...