# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ace Digital OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageAttachment } from './messageAttachment';
import type { MessageInputMessageKind } from './messageInputMessageKind';
import type { MessageInputMetadata } from './messageInputMetadata';

export interface MessageInput {
  body?: string;
  attachments?: MessageAttachment[];
  messageKind?: MessageInputMessageKind;
  metadata?: MessageInputMetadata;
}